Generate KeyAccid attribute children
